+++ engine/src/main/java/net/jqwik/engine/facades/ArbitrariesFacadeImpl.java
@@ -83,7 +83,7 @@ public class ArbitrariesFacadeImpl extends Arbitraries.ArbitrariesFacade {
 	}
 
 	@Override
-	public <M> ActionSequenceArbitrary<M> sequences(Arbitrary<Action<M>> actionArbitrary) {
+	public <M> ActionSequenceArbitrary<M> sequences(Arbitrary<? extends Action<M>> actionArbitrary) {
 		return new DefaultActionSequenceArbitrary<>(actionArbitrary);
 	}