+++ docs/release-notes.md
@@ -26,6 +26,9 @@ title: jqwik Release Notes
 
 #### New and Enhanced Features
 
+- `Arbitraries.sequences(Arbitrary<? extends Action<M>> actionArbitrary)` does now accept
+  covariant subtypes and still returns type `Arbitrary<Action<M>>`.
+
 - Added `ActionSequence.peek(Consumer<M> modelPeeker)`.
 
 - Reporting text on failed statistics coverage check improved.